feat: Add redis pool size (#2727)
## Context

We aim to make the Redis connection pool size for the cache configurable
via an environment variable
The default Redis connection pool size is currently set to 5

## Description

This PR introduces a new environment variable
`LAGO_REDIS_CACHE_POOL_SIZE` to control the Redis connection pool size
for caching. If the environment variable is set, the value will be used
as the pool size; if not, it will default to 5.