misc: Avoid reloading charge filter if already loaded (#2742)
## Context

This PR is related to https://github.com/getlago/lago-api/pull/2678.
The charge filter lookup that was added in the
`Events::PostProcessService` lead to some performance regression. The
reason behind this is that the `ChargeFilters::EventMatchingService` is
reloading systematically the charge filters and associated values even
if they were already pre-loaded in the `Events::PostProcessService`

## Description

The PR is checking the cached status of the `filters` association to
avoid the reload.