feat(netsuite): Cleanup syncs (#2782)
## Context

When a users searches for items to map the UI is not displaying all the
items. There is a reason for this
is that the NetSuite SuiteScript limits to 1000 results.

## Description

This PR we change the way we map the items. Instead of selecting the
items from a prefetched list, the user needs to manually fill out the
name, code, and the account id.