feat(salesforce): Added customers webhooks (#2783)
## Context

This PR adds to the Customers::CreateService and
Customers::UpdateService the SendWebhookJob with appropriate events
('customer.created' and 'customer.updated'). This ensures that both
customer creation and updates send the needed webhooks
## Description
`Customers::CreateService`
Added SendWebhookJob.perform_later('customer.created', customer) to the
end of the create_from_api and create methods, so a webhook is sent each
time a new customer is created.

`Customers::UpdateService`
Added SendWebhookJob.perform_later('customer.updated', customer) to the
end of the call method, ensuring that updates to customer send a webhook
event.