bug(DataExport) - Decrease batch size for data export & use tempfile (#2786)
## Description

when running sidekiq with high amount of concurrency, the default batch
size could result in large amounts of longer running jobs. These could
make a worker consume too much memory.

This PR aims to fine-tune the batch size by reducing it by a factor 5
(so jobs will be faster and consume less memory individually). It will
also not keep the entire CSV in memory, but rather use a tempfile.