feat(plan-charges-cascading): Support charge filters cascading (#2761)
## Context

Currently when plan is updated, these changes are not cascaded to all
the children plans.

## Description

This PR supports cascading of charge filters. There are several rules:

- If main attribute that supports cascading is not true -> cascading is
disabled
- Charge filter properties are cascaded only if the same properties are
not already overridden in the child filter
- We need to ensure that all filters that are specific to the child are
not changed in any way (extra filters that don't exist on parent)
- When performing cascading, we are basically at the same time updating
parent filters and children filters. However, we need to ensure that
logic which handles cascading on children filter uses parent filter
attributes **before** update for comparison. That's why we pass parent
filter attributes before any update to the children logic.

---------

Co-authored-by: Anna Velentsevich <annvelents@gmail.com>