add migration to set parent_ids to charges (#2738)
## Context

We want to populate parent_ids for charges. To define a parent for a
charge, we're looking into plan's parent plan and trying to find a
charge with matching charge_model and properties. If parent plan has
more than one matching charge, we claim that we're not able to define a
parent charge

## Description

added a migration that:
- goes through children plans and loads their charges and parents with
their charges
- for each charge of children plan it tries to select charge from parent
plan with matching charge model and charge properties
- if only one matching charge is found, update child charge with parent
charge (without callbacks)

added a rake task to seed test data, so we can deploy it onto staging,
create any number of plans with children, select to not save parent_ids
for their charges, run the migration code and check the result