misc(stripe): Refact payment method update webhooks (#2808)
## Context

This PR is related to the remaining `
PaymentProviders::Stripe::HandleEventJob` dead jobs with
`BaseService::NotFoundFailure: stripe_customer_not_found` error

## Description

The main goal of the PR is to refactor the handling of the
`setup_intent.succeeded` and `customer.updated` to move the logic in a
dedicated webhook service (in the objective of killing the
unmaintainable `PaymentProviderCustomers::StripeService`).

It also fixes a case in `setup_intent.succeeded` when the webhook has a
`customer` with null value and no Lago ids in the metadata. Since in
this specif case the webhook does not target a lago resource, the
handling should not raise an error but should be ignored.