fix(stripe): Handle missing payment (#2825)
## Context

Some `PaymentProviders::Stripe::HandleEventJob` with
'payment_intent.payment_failed' or 'payment_intent.succeeded' event type
are failing with `BaseService::NotFoundFailure -
stripe_payment_not_found`.

The webhook contains metadata of lago's invoices but the related
`Payment` cannot be found in the database. It is probably because the
job creating the stripe payment was killed allong with the worker
running it under heavy load.

## Description

This PR changes the logic to re-create the `Payment` record when missing
using the data from the webhook object. It will allow to update the
`invoice.payment_status` accordingly