feat(daily_usage): Compute and store usage_diff (#2800)
## Context

This PR is part of the Usage Revenue and unit.

Today, Lago does not offer a way to retrieve customer usage with a
granularity lower than the billing period (via invoices and fees). On
the other end, it is possible to get the current usage for a
customer/subscription, but this usage is just a “snapshot” of the usage
a the current time.

## Description

This PR adds a new `usage_diff` field to the `DailyUsage`. The logic to
compute the daily usage is also updated to fill this new field.