feat(dunning): Add DunningCampaign resolver to graphql api (#2834)
## Roadmap Task

👉 https://getlago.canny.io/feature-requests/p/set-up-payment-retry-logic

👉 https://getlago.canny.io/feature-requests/p/send-reminders-for-overdue-invoices

 ## Context

We want to automate dunning process so that our users don't have to look
at each customer to maximize their chances of being paid retrying
payments of overdue balances and sending email reminders.

We are extending dunning campaigns management to edit and delete
campaigns.

 ## Description

This change adds a new resolver to graphql api to retrieve a single
dunning campaign by id. This is required to render the dunning campaign
edit screen in the frontend side.

Also adds counter cache customers_count on dunning campaigns customer
relationship to provide a performant count of customers with the dunning
campaign applied.