feat(dunning): reset customer dunning counters on invoice updates
 ## Roadmap

👉 https://getlago.canny.io/feature-requests/p/set-up-payment-retry-logic

👉 https://getlago.canny.io/feature-requests/p/send-reminders-for-overdue-invoices

 ## Context

We want to automate dunning process so that our users don't have to look
at each customer to maximize their chances of being paid retrying
payments of overdue balances and sending email reminders.

We are extending dunning campaigns management to edit and delete
campaigns.

 ## Description

When an invoice included in a payment request that belongs to a dunning
campaign is payment succeeded, this change resets the customer dunning
campaign status counters (completed, last attempt number and last attempt timestamp)