fix (grace_period): refactor how issuing date is set (#2928)
## Context

When billing job is triggered in the past, `issuing_date` is not set
correctly.

## Description

`issuing_date` is set based on the billing job timestamp and grace
period.

However, when we check if it is correct time to finalize invoice or in
services that set grace period we calculate issuing date again based on
`invoice.created_at`. If `invoice.created_at` is different than billing
day timestamp, we could end-up with invalid issuing date. This
especially affects when the draft invoice is finalized.

@vincent-pochet @rsempe Can you see any downside of this approach?