fix: Ensure Plans::DestroyJob are queued only once (#3010)
This change is meant to avoid issues when a user tries to delete a plan
more than once. This happens because the plan deletion is performed as a
background job and could take some time as all child plans must be
deleted, subscriptions terminated, invoices generated and discarded,
etc...

The goal of this change is to avoid failed jobs because of race
conditions between multiple runs for the same plan producing unique DB
index constraints errors or discard! errors as those only produce noise
and waste of support time.