fix: Recurring fee not displayed on invoices after upgrade (#3012)
## Context

Steps to reproduce the issue:

1. Create a recurring metric;

2. Create a plan with a charge linked to the recurring metric - the
charge should be:
* prorated = true;
* pay_in_advance = true;
* invoiceable = false;
* regroup_paid_fees = invoice;

3. Create a subscription with a start date in the past and send an event
in the past (e.g. 1 December) -> you receive a fee.created webhook;

4. Duplicate the original plan and only modify the price of the charge
(higher price);

5. Trigger an upgrade of the existing subscription by assigning the new
plan;

6. Mark the fee in step 3 as succeeded via POST api/v1/fees/:lago_id

7. Terminate the subscription -> the fee is not included in the
end-of-period invoice.

Expected behaviour: the fee should be included in the end-of-period
invoice, even if the original Lago subscription to which it belongs is
no longer active.

## Description

This change fetches all subscriptions with status active or terminated
of a given organization with a matching external id to invoice their
paid in advance fees.