misc(wallet): Refactor wallet transaction related jobs queuing (#3040)
## Context
Queuing of wallet transaction related jobs is ideal as jobs as a low
priority, meaning that under heavy load, they are a huge delay might
happen between enqueue and processing. Since it is user facing, we need
to improve it.


## Description

This PR:
- Adds a new `high_priority` queue that should be used for all jobs
requiring an "immediate" processing
- Moves the `Invoices::PrepaidCreditJob` from the `wallet` to the new
`high_priority` queue
- Moves the `WalletTransactions::CreateJob` from the `wallet` to the new
`high_priority` queue
- Moves `Wallets::RefreshOngoingBalanceJob` into the low priority queue

Note no more jobs will relies on the `wallet` queue, but this queue is
kept fro now in the config to avoid issue with existing enqueued jobs.