fix(stripe): Payment failing with requires_payment_method should be marked as failed (#3042)
## Context

Some recent refactors were made on the payment processing implementation
with stripe.

When a payment intent creation was failing with the
`requires_payment_method` status, the lago payment is kept `processing`
preventing it from being retried because of the idem-potency key (the
new attempt will have a different payload if the payment method is
updated).

## Description

This PR changes the behavior for `requires_payment_method` ensuring that
the payment is marked as failed, allowing a retry after fixing the cause
of the failure (missing or expired payment method)