Propagate StaleObjectErrors so PayInAdvance jobs retry (#3036)
## Context

Pay in advance jobs do not reliably retry when a StaleObjectError is
raised.

## Description

This PR builds on top of [this
PR](https://github.com/getlago/lago-api/pull/3027) which will retry the
pay in advance job when a StaleObjectError is raised. This error is
currently handled by the generic `rescue` block, and instead should be
re-raised similarly to how the SequenceError is.

I have tested this in a local docker setup, by sending a large volume of
events to Lago over a very short period. Before this fix, some fees fail
to generate, after the fix all the fees are successfully generated.