misc(PaymentRequest): Apply payment idempotency refactor (#2986)
## Context

This PR follow https://github.com/getlago/lago-api/pull/2962 in the
refactor of the payment processing. It applied the logic that was
applied to payment for invoices to payment requests

The main goal of this refactor will be to:
- Avoid double payment at all cost
- Avoid code duplication between payment processor integration to make
maintenance easier

## Description

This PR applies the following logic for payment processing of payment
requests:
- Creates a payment (or retrieve the pending one in case of retry) 
- Delegate the payment processing to the dedicated
`PaymentProviders::*::Payments::CreateService` responsible only for the
payment creation and the error handling (including idempotency). The
payment provider payment creation logic will be shared between types of
payment and will rely only on the Lago Payment record