Feat manual payments controller (#3045)
## Roadmap Task

👉  https://getlago.canny.io/feature-requests/p/log-partial-payments

## Context

This PR introduces a new endpoint to list all payments. Additionally, it
adds the ability to filter payments by a specific invoice_id, allowing
users to narrow down the results to payments associated with a
particular invoice. This enhancement improves the flexibility and
usability of the payments listing functionality.
It also adds `create` and `show` actions in the PaymentsController.

## Description

**Added Endpoint:**

- Created an endpoint to list, show and create payments.
- Filter by invoice_id:
- Added functionality to filter payments based on the provided
invoice_id.
- Included a validation step to ensure the invoice_id is in the correct
UUID format before executing the query.

**Query Updates:**

- Refactored PaymentsQuery to handle dynamic filtering by invoice_id.
- Ensured that the query properly joins related tables, such as invoices
and payment_requests, to support filtering.

---------

Co-authored-by: brunomiguelpinto <darkymiguel@gmail.com>