feat(invoice-preview): add base preview service (#3056)
## Context

Preview feature enables fetching first Lago invoice. This invoice is not
persisted aand is calculated on the fly.

## Description

This PR adds preview service which will be the base for the feature.
It can work for existing subscription and customer, but also for objects
that are just created in-memory.

Following PRs will attach credits to the preview invoice and tax
provider integration will also be supported.