fix(stripe): Idem-potency errors should keep the payment pending (#3063)
## Description

This PR is a fix related to the recent payment handling refactor.

It makes sure that payment attempt on stripe failing with an
`Stripe::IdempotencyError` keeps a `pending` payable payment status.
It will fix the retry handling of payment failed with an
`amount_too_small` error message when a customer payment method is
updated