feat(invoice-preview): add logic for applying coupons on preview invoice (#3073)
## Context

Preview feature enables fetching first Lago invoice. This invoice is not
persisted and is calculated on the fly.

## Description

This PR adds service for applying coupons on preview invoice. The goal
was not to affect heavily main service for applying coupons on invoices.

In order not to repeat the code, some common logic is extracted s that
it can be used on both places