fix(cashfree): Fix webhook and payment flow (#3106)
## Description

Updated the Cashfree webhook integration to handle both sets of headers.
Additionally, added `payment_type: "one-time"` to the link parameters to
ensure the payment is created upon receiving the webhook.

## Additional Context
I tested the flow with these changes in the actual Cashfree Test
Environment, and everything is now functioning as expected.

fixes #3105