Reapply "feat(invoice): Avoid generating 0 amount charge fees (#3032)" (#3059) (#3068)
## Description

This PR follow https://github.com/getlago/lago-api/pull/3020 the goal is
to stop generating empty fees (0 units, amount and events without
adjustments or true up fees).

The reason behind this change is scalability as for now, depending on
the plan setups, it appears that 98% of generated fees does not have an
amount...

**NOTE:** a new feature will improve the edit fees feature since this PR
will prevent creation of adjustment fees from the UI if no event have
been received for a specific billable metric on a period.