feat(stripe): Save Payment Method in Customer when paying invoice with checkout url (#3107)
## Context

Recently we added the possibility to pay an invoice with a stripe
checkout url. The payment method used wasn't saved in stripe. This
change will allow the customer to reuse this method next time.


## Description

When using the `/customers/:external_customer_id/checkout_url` you setup
a payment method so the payment method is associated with the customer
in Stripe.

When using `/invoices/:lago_id/payment_url` the payment method is used
but not associated. This PR makes sure the payment method used is savec
with the customer.

Related docs:
https://docs.stripe.com/api/checkout/sessions/create#create_checkout_session-payment_intent_data-setup_future_usage

For now, we don't support offsession paiement in India so we don't et
this parameter in case the customer is in India.

Note that **it's different** from saving your card with Link (as shown
below).

![CleanShot 2025-01-27 at 15 12
32@2x](https://github.com/user-attachments/assets/1469f261-cd3f-43c3-9ca8-451c4a77d993)