misc(adjusted_fee): Allow fee adjustement without pre-existing fee (#3083)
## Context

As a consequence of the removal of the 0 units fees removal from both DB
and invoicing logic, we lost the ability to adjust fees without amount /
units as the feature was expecting a fee to be present. This feature is
a new approach to bring this functionality back.

## Description

This features refactors the `AdjustedFees::CreateService` and the
`Mutations::AdjustedFees::Create` mutation to allow the creation of
adjusted fees without pre-existing fees via the GraphQL API