misc: Introduce BaseResult to get rid of OpenStruct (#3122)
## Context

As a requirement to enable YJIT compiler for Ruby, we need to get rid of
OpenStruct as it leads to a huge memory consumption.

At the same time, we are lacking visibility on what properties are
attached to result objects returned by services.

## Description

This PR introduces a new `BaseResult` result object, that should be used
a a base in every services to define the attributes of a Result:
```ruby
module SomeModule
  class SomeService < BaseService
    Result = BaseResult[:attr_name1, :attr_name2]
```

To keep compatibility until all services are migrated to this approach,
the `BaseService` is still relying on a `LegacyResult` using OpenStruct
as a parent class.