fix: clickhouse related specs (#3138)
## Context
This PR fixes issues in specs and migration specs that were introduce by
the recent merge of https://github.com/ClickHouse/ClickHouse/pull/75259
on clickhouse side.

## Description

The network access for the `default` user is now disabled when no
password is provided.
This PR makes sure that the password is effectively provided