fix(anrok): Remove cache for Current Usage (#3142)
## Context

When calling `GET
/customers/:cust_id/current_usage?external_subscription_id=:sub_id` Lago
computes the current usage

## Description

The taxes are computed by a thrid party (Anrok through Nango). We
originally implemented a cache to avoir http calls to third party but it
comes with drawbacks. We recently [tried to improve
this](https://github.com/getlago/lago-api/pull/3115).
After investigation we believe it's best to remove the cache.

- [x] Remove Anrok cache
- [x] Introduce new `item_key` to reconcile fees when they are not
persisted
- [x] Add new `apply_taxes` params to the API
- If you get an error due to third party, you can call again without
taxes
- In case you only need pre-tax amounts (like for b2b), passing
`apply_taxes: false` will be more performant
  - https://github.com/getlago/lago-openapi/pull/332

---------

Co-authored-by: Miguel Pinto <darkymiguel@gmail.com>