fix(invoice): Add GraphQL flag to know editable status (#3152)
## Context

An issue was reported with creation of adjusted fees following the
recent changes related to the removal of zero amount fees from draft
invoices.

In case of `subscription_starting` invoice_subscription, adjustment
creation is ignored because the invoice refresh is not creating charge
fees. This behavior is expected as no charge fees are created when
starting a subscription.
Updating the back-end to allow such edition would require a lot of
changes with high risk of unexpected side effect.

## Description

This PR adds a new `AcceptNewChargeFees` boolean flag to the
`InvoiceSubscription` GraphQL object to let the front-end know if the
add new fee button should be displayed or not