feat(api_key): cache api key and organization (#3137)
## Context

This PR is part of a global initiative to improve the performance of the
LAGO Rest API.

Today every request made received by the API leads to at least two
database queries to authenticate the organization. A first one is
fetching the `api_key` using the provided bearer token from the HTTP
header, the second one loads the related `organization`.

As by nature, api_keys and organizations are not update really often,
adding a caching strategy could reduce drastically the time consumed
authenticating the organization.

## Description

This PR changes the authentication mechanism by:
- Adding a dedicated cache service for API key. It default to no cache
unless cache is explicitly turned on.
- If cache is on:
  - Checking for cached api_key and organization
  - Fetching the values from the DB is no cache is present
  - Caching the values
  - Handling cache expiration when updating api keys and organization