chore(spec): group scattered it blocks for Graphql Types (#3167)
## Context

Our test suite has become pretty long to run.

## Description

This change **saves 951 examples**.

For all GQL types, we use one `it` block per line. It looks very good
but this is not ideal. This PR groups all the it blocks

> [!NOTE]
Once merged, I'll add the merge commit to the `.git-blame-ignore-revs`
file.

**Before**

```ruby
  it { is_expected.to ... }
  it { is_expected.to ... }
  it { is_expected.to ... }
```

**After**

```ruby
  it do
    expect(subject).to ...
    expect(subject).to ...
    expect(subject).to ...
  end
```

![CleanShot 2025-02-11 at 12 32
43@2x](https://github.com/user-attachments/assets/cf029987-221c-4340-a948-2dbca8dee6d8)