feat(wallet): New `wallet.(created|updated|terminated)` webhooks (#3133)
## Context

Wallet operations trigger 3 webhooks today: 
* `wallet.depleted_ongoing_balance`
* `wallet_transaction.updated`
* `wallet_transaction.payment_failure`.

We want to add the following: 
* `wallet.created`
* `wallet.updated`
* `wallet.terminated`


https://docs.getlago.com/api-reference/webhooks/messages#wallets-and-prepaid-credits

## Description

### `wallet.created`

~Notice that the webhook job is dispatched by
`WalletTransactions::CreateJob` so the wallet sent in the webhook has
the initial balance instead of 0. Unlike the API which returns 0.~
~The `WalletTransactions::CreateJob` is always dispatch even if the
`WalletTransactions::CreateService` might do nothing.~

Notice that the webhook job is dispatched as soon as the wallet is
created so the initial balance can be 0. The API always returns balance
= 0 when creating a wallet.

The initial balance is set async by the `WalletTransactions::CreateJob`
job and it runs in the `high_priority` queue.

> [!NOTE]
Creating a wallet transaction to set the initial balance will also send
a `wallet.updated` webhook. It's possible to receive this one **before**
the `wallet.created`.

We still send the wallet.created because you can create a wallet without
initial transaction but **we should never rely on the order**.

The `WalletTransactions::CreateJob` is always dispatch even if the
`WalletTransactions::CreateService` might do nothing.

### `wallet.terminated`

The webhook is set only of the wallet just got terminated. If the
service is called on a terminated service, no webhook is sent.

### `wallet.updated`

This webhook is dispatched when 
* an attribute of the wallet is updated (name, expiration date or
invoice_requires_successful_payment config)
* a recurring transaction rule is updated because recurring **rules are
part of a wallet**, unlike wallet_transactions which are "related to a
wallet" (and have their own webhooks)
* the balance is increased or decreased

**No webhook is sent when the ongoing balance is updated** because it
would trigger too many events. Tracking ongoing balance is part of the
"real time usage" effort.

---------

Co-authored-by: Vincent Pochet <vincent@getlago.com>