chore(rubocop): Disable or fix last errors (#3179)
As I'm fixing the quotes for string literals, there are other errors
showing up so let's fix those. I'm going to merge this one first before
updating the quote PRs.

Many of them are about old migrations that I don't think we should
change. I simply disabled the rule for them.

`Cops/ServiceCallCop` was renamed `Lago/ServiceCall` and I excluded the
services failing. It's not so easy to fix in this kind of cosmetic PR
but it should be done one day.

> 3234 files inspected, no offenses detected

![CleanShot 2025-02-13 at 13 58
18@2x](https://github.com/user-attachments/assets/1ec3f073-152a-4219-b2fe-165b1565a983)