chore(style): Fix all string litteral quotes (Style/StringLiterals) (#3177)
## Description

We want to start using double quotes for string litteral to follow the
standard-rb convention.

Rather than fixing them in many many PR and pollution the `git blame`
we'll fix all the codebase in on PR and add the final commit to the
`.git-blame-ignore-revs` file (so I disappear from the blame!). The
rubocop rule will be enabled in the second PR too because I want this
one to have **only** quote changes.


https://github.blog/changelog/2022-03-24-ignore-commits-in-the-blame-view-beta/

This will create tons of conflict with open PR. I'm happy to help rebase
🙏