feat(regroup_fee): Sum fees to get invoice totals instead of recomputing taxes (#3153)
## Context

Using Fees with the following attributes, you will get an invoice at the
end of the period that regroups all PAID fees.
* pay_in_advance: true
* invoiceable: false
* regroup_paid_fees: 'invoice'

## Description

The taxes on these fees are computed for each fee. The final invoice
should show the amount of taxes actually paid, not the amount of taxes
that would have been paid if the fees were paid at once in a single
invoice invoice.

**Example**

Fee 1:
* amount_cents: 12
* taxes: 20%
* taxes_amount_cents: 2.4 rounded to 2

Fee 2:
* amount_cents: 12
* taxes: 20%
* taxes_amount_cents: 2.4 rounded to 2

Final invoice:
* 2 fees 
* fees_amount_cents: 24 (12 * 2)
* taxes_amount_cents: **4 if you sum the existing fees, but 20% of 24 is
4.8, rounded to 5**
* total_amount_cents: 28 with this PR (actually paid taxes) but 29
before. There is 1 cent invoiced but not collected.