fix(stripe): Avoid raising not_found for deleted stripe customers (#3194)
## Context

Some `customer.updated` Stripe webhooks are failing with a
`stripe_customer_not_found` error.

The metadata of the webhook payload includes a `lago_customer_id`,
meaning that it was initially created from Lago and should be found in
the database.

After some investigation it appears that a deleted `StripeCustomer` is
present in the database meaning the the webhook should not raise an
error but should instead be ignored.

## Description

This PR checks for the presence of a deleted `stripe_customer` to return
an empty success result when a deleted one is found in the database.

---------

Co-authored-by: Romain Sempé <romain@getlago.com>