fix(payment): Handle deleted customers (#3196)
## Context

When we delete a customer, all related subscriptions are terminated and
a termination invoice is generated. If the customer is linked to a
payment provider, the system will try to collect it.

Some error will be raised as the application will not be able to find
the discarded record in the database

## Description

This PR extract the
`PaymentProviderCustomers::StripeService#check_payment_method` into a
dedicated `PaymentProviderCustomers::CheckPaymentMethodService` and
makes sure it relies only on the `stripe_customer` to fetch the payment
method.