misc(mailer): Improve attachements names (#3201)
## Context

Currently, invoices and credit notes attached to emails are named
generically (e.g., invoice.pdf, credit_note.pdf). This makes it
difficult for customers to organize and retrieve specific documents
especially when dealing with multiple invoices or credit notes.

## Description

This PR improves the naming by including the number in the attachments
file names:
- For invoices: `invoice-NUMBER.pdf`
- For credit notes: `credit_notes-NUMBER.pdf`