fix: Add unique index on payment provider id (#3223)
## Context

We have seen some duplicate payments due to a race condition when
receiving two stripe webhooks at the same time.

## Description

The objective of this PR is to add a unique constraint for the table
`payments`, on the columns `provider_payment_id` and
`payment_provider_id`.

We took the opportunity to fix invalid `total_paid_amount_cents` on the
impacted invoices.