misc: Update rails to v7.2 (#2894)
## Context

In order to update Rails to v8.0 first we update its version to the next
and latest minor rails v7 version (v7.2.x)

## Description

* `rails` is updated from `7.1.2.4` to `7.2.2`

### Impact

all tests now respect the `active_job.queue_adapter` config
[link](https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#all-tests-now-respect-the-active-job-queue-adapter-config)

rails configuration files updated.

Ruby YJIT is disabled as we are not ready because of the usage of
OpenStruct for service result responses.

APP is live in: https://rails72-app.staging.getlago.com/

---------

Co-authored-by: Vincent Pochet <vincent@getlago.com>